fix(travis): rename file wit a letter deletion -[Fixes #bg-fork-travis-build-164583983]
