+++ server/controllers/MessageController.js
@@ -1,5 +1,6 @@
+
 import messages from '../models/messages';
-import Validate from '../validators/Validate';
+import Validate from '../validators/Validates';
 
 /**
  * Class representing API endpoints for
