+++ server/routes/messages.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import MessagesController from '../controllers/MessagesController';
+import MessagesController from '../controllers/MessageController';
 
 const router = express.Router();
